split main.go into functionally separate files, add some logs
