+++ hash_test.go
@@ -27,25 +27,6 @@ func TestLongPassHash(t *testing.T) {
     }
 }
 
-func TestAverage(t *testing.T) {
-    // first clear the duration map
-    for k := range durationMap {
-        delete(durationMap, k)
-    }
-
-    // add some durations
-    durationMap[0] = 1012; incrementCounter()
-    durationMap[1] = 2034; incrementCounter()
-    durationMap[2] = 536; incrementCounter()
-    durationMap[3] = 9876; incrementCounter()
-
-    rc := average()
-    expected := (durationMap[0] + durationMap[1] + durationMap[2] + durationMap[3]) / 4
-    if rc != expected {
-        t.Errorf("Average was incorrect, got: %d, want: %d.", rc, expected)
-    }
-}
-
 func TestId(t *testing.T) {
     validateId(t, "/hash/34", 34)
     validateId(t, "/hash/", 0)
@@ -58,6 +39,6 @@ func validateId(t *testing.T, path string, expected int64) {
     id := id(path)
 
     if id != expected {
-        t.Errorf("Average was incorrect, got: %d, want: %d.", id, expected)
+        t.Errorf("Extracted id was incorrect, got: %d, want: %d.", id, expected)
     }
 }
